Attempt to move uptime into components folder
